+++ scripts/bundle.js
@@ -7,7 +7,16 @@ const { mkdirSync, writeFileSync, copyFileSync } = require('node:fs');
 async function bundle() {
   mkdirSync('dist', { recursive: true });
 
-  copyFileSync('src/preload.cjs', 'dist/preload.js');
+  await esbuild.build({
+    entryPoints: ['src/preload.cjs'],
+    bundle: true,
+    minify: true,
+    platform: 'node',
+    external: ['mitt','fs','path','os','crypto','util'],
+    target: ['node16'],
+    outfile: 'dist/preload.js',
+    logLevel: 'info'
+  });
 
   await esbuild.build({
       entryPoints: ['src/renderer/renderer.js'],